requirements.txt
streamlit
google-generative-ai
youtube_transcript_api
python-dotenv
pathlib